requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg[binary]==3.1.18
python-dateutil==2.8.2
django-filter==23.5
drf-spectacular==0.27.0